[bug]: fix bugs in new the schema generation process
