+++ docs/source/references/api_spec.md
@@ -12,7 +12,7 @@
 
 **Method**: `POST`
 
-**Expects**: Body: [`VersionedSignedTransaction`](#iroha-structures)
+**Expects**: Body: `VersionedSignedTransaction`
 
 **Responses**:
 
@@ -34,7 +34,7 @@
 
 **Expects**:
 
-- Body: [`VersionedSignedQuery`](#iroha-structures)
+- Body: `VersionedSignedQuery`
 - Query parameters:
     - `start`: Optional parameter in queries where results can be indexed. Use to return results from specified point.
       Results are ordered where can be by id which uses
@@ -45,7 +45,7 @@
 
 **Responses**:
 
-| Response        | Status | [Body](#iroha-structures)                     |
+| Response        | Status |                      Body                     |
 |-----------------|--------|-----------------------------------------------|
 | Signature err.  | 401    | `QueryExecutionFailure::Signature(String)`    |
 | Permission err. | 403    | `QueryExecutionFailure::Permission(String)`   |
@@ -86,8 +86,7 @@ Whether each prerequisite object was found and `FindError`:
 
 **Communication**:
 
-After handshake, client should send [`VersionedEventSubscriptionRequest`](#iroha-structures). Then server
-sends [`VersionedEventMessage`](#iroha-structures).
+After handshake, client should send `VersionedEventSubscriptionRequest`. Then server sends `VersionedEventMessage`.
 
 **Notes**:
 
@@ -130,8 +129,7 @@ _Internal use only._ Returns the transactions pending at the moment.
 
 **Communication**:
 
-Client should send [`VersionedBlockSubscriptionRequest`](#iroha-structures) to initiate communication after WebSocket
-handshake. Then server sends [`VersionedBlockMessage`](#iroha-structures).
+Client should send `VersionedBlockSubscriptionRequest` to initiate communication after WebSocket handshake. Then server sends `VersionedBlockMessage`.
 
 **Notes**:
 
@@ -381,17 +379,3 @@ codec's [GitHub repository](https://github.com/paritytech/parity-scale-codec).
 ## Reference Iroha Client Implementation
 
 [Iroha client in Rust.](../../../client)
-
-## Iroha Structures
-
-| Object name                         | Fully qualified object name                                    |
-|-------------------------------------|----------------------------------------------------------------|
-| `VersionedSignedTransaction`        | `iroha_data_model::transaction::VersionedSignedTransaction`    |
-| `VersionedSignedQuery`              | `iroha_data_model::query::VersionedSignedQuery`                |
-| `VersionedPaginatedQueryResult`     | `iroha_data_model::query::VersionedPaginatedQueryResult`       |
-| `QueryExecutionFailure`             | `iroha_core::smartcontracts::isi::query::Error`                |
-| `FindError`                         | `iroha_core::smartcontracts::isi::error::FindError`            |
-| `VersionedBlockSubscriptionRequest` | `iroha_core::block::stream::VersionedBlockSubscriptionRequest` |
-| `VersionedBlockMessage`             | `iroha_core::block::stream::VersionedBlockMessage`             |
-| `VersionedEventSubscriptionRequest` | `iroha_data_model::events::VersionedEventSubscriptionRequest`  |
-| `VersionedEventMessage`             | `iroha_data_model::events::VersionedEventMessage`              |
