+++ schema/derive/tests/ui_fail/fieldless_enum_with_explicit_discriminant.rs
@@ -1,0 +1,11 @@
+use iroha_schema::IntoSchema;
+
+#[derive(IntoSchema)]
+enum EnumWithExplicitDiscriminant {
+    A = 1,
+    B,
+    C,
+    D,
+}
+
+fn main() {}
