+++ schema/tests/fieldless_enum.rs
