+++ tools/parity_scale_decoder/Cargo.toml
@@ -16,15 +16,14 @@ no-color = ["colored/no-color"]
 iroha_data_model = { version = "=2.0.0-pre-rc.13", path = "../../data_model", features = ["http"]}
 iroha_primitives = { version = "=2.0.0-pre-rc.13", path = "../../primitives", default-features = false }
 iroha_schema = { version = "=2.0.0-pre-rc.13", path = "../../schema"}
+iroha_schema_gen = { version = "=2.0.0-pre-rc.13", path = "../../schema/gen"}
 iroha_crypto = { version = "=2.0.0-pre-rc.13", path = "../../crypto", default-features = false }
+iroha_version = { version = "=2.0.0-pre-rc.13", path = "../../version", default-features = false }
 iroha_genesis = { version = "=2.0.0-pre-rc.13", path = "../../genesis"}
 clap = { version = "3.2.23", features = ["derive", "cargo"] }
 eyre = "0.6.8"
 parity-scale-codec = { version = "3.2.1", default-features = false }
 colored = "2.0.0"
-
-[dev-dependencies]
-iroha_schema_gen = { version = "=2.0.0-pre-rc.13", path = "../../schema/gen"}
 
 [build-dependencies]
 iroha_data_model = { version = "=2.0.0-pre-rc.13", path = "../../data_model"}
